agents/market_research_agent.py
import requests
import json
from bs4 import BeautifulSoup
from typing import List, Dict
from .base_agent import BaseAgent
from models.schemas import BusinessInput, MarketResearchResult

class MarketResearchAgent(BaseAgent):
    def __init__(self):
        super().__init__("MarketResearch")
    
    async def execute(self, business_input: BusinessInput) -> MarketResearchResult:
        # Simulate market research (in production, use real APIs)
        trending_keywords = await self._generate_trending_keywords(business_input)
        competitor_insights = await self._analyze_competitors(business_input)
        search_data = await self._get_search_volume_data(trending_keywords)
        
        return MarketResearchResult(
            trending_keywords=trending_keywords,
            competitor_insights=competitor_insights,
            search_volume_data=search_data,
            difficulty_scores={kw: 0.5 for kw in trending_keywords}
        )
    
    async def _generate_trending_keywords(self, business_input: BusinessInput) -> List[str]:
        system_prompt = """You are a market research expert. Generate trending keywords 
        related to the given business information. Focus on long-tail keywords with 
        commercial intent."""
        
        user_prompt = f"""
        Business Type: {business_input.business_type}
        Product/Service: {business_input.product_service}
        Target Audience: {business_input.target_audience}
        Niche Keywords: {', '.join(business_input.niche_keywords)}
        
        Generate 10 trending keywords that would be valuable for SEO content.
        Return as a comma-separated list.
        """
        
        result = await self._call_llm(system_prompt, user_prompt)
        return [kw.strip() for kw in result.split(',')]
    
    async def _analyze_competitors(self, business_input: BusinessInput) -> List[Dict[str, str]]:
        # Simplified competitor analysis
        system_prompt = """You are analyzing competitors for SEO strategy. 
        Provide insights about what competitors might be doing well."""
        
        user_prompt = f"""
        Business: {business_input.business_type} - {business_input.product_service}
        
        Provide 3 competitor insights in the format:
        1. Strategy: [strategy description]
        2. Content Gap: [content opportunity]
        3. Keyword Focus: [keyword strategy]
        """
        
        result = await self._call_llm(system_prompt, user_prompt)
        
        # Parse the result into structured format
        lines = result.split('\n')
        insights = []
        for i, line in enumerate(lines[:3]):
            insights.append({
                "insight_type": ["strategy", "content_gap", "keyword_focus"][i],
                "description": line.strip()
            })
        
        return insights
    
    async def _get_search_volume_data(self, keywords: List[str]) -> Dict[str, int]:
        # Simulate search volume data (in production, use real SEO APIs)
        import random
        return {kw: random.randint(100, 5000) for kw in keywords}