requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.0
jinja2==3.1.2
textstat==0.7.3
streamlit==1.28.0
pandas==2.1.0
plotly==5.17.0